+++ CHANGELOG.md
@@ -12,6 +12,7 @@ of tasks being currently executed.
 - JAMES-2904 Authentication and SSL support for Cassandra backend
 - JAMES-2904 Authentication and SSL support for ElasticSearch backend
 - JAMES-3066 Add "allowed From headers" webadmin endpoint
+- JAMES-3062 EventDeadLettersHealthCheck
 
 ### Changed
 - Multiple changes have been made to enhance ElasticSearch performance: